// webui/src/sdk/models/JobType.ts
export type JobType =
  | "hero"
  | "darkknight"
  | "paladin"
  | "bishop"
  | "archmagefb"
  | "archmagetc"
  | "nightlord"
  | "shadower"
  | "dualblade"
  | "bowmaster"
  | "sniper"
  | "pathfinder"
  | "corsair"
  | "buccaneer"
  | "cannoneer"
  | "soulmaster"
  | "flamewizard"
  | "windbreaker"
  | "nightwalker"
  | "striker"
  | "mihile"
  | "demonavenger"
  | "demonslayer"
  | "battlemage"
  | "wildhunter"
  | "mechanic"
  | "zenon"
  | "blaster"
  | "evan"
  | "luminous"
  | "mercedes"
  | "phantom"
  | "eunwol"
  | "aran"
  | "kaiser"
  | "kain"
  | "cadena"
  | "angelicbuster"
  | "adele"
  | "illium"
  | "ark"
  | "lara"
  | "hoyoung"
  | "zero"
  | "kinesis"
  | "virtual_maplestory_m";